requirements.txt
Django==1.11.8
gunicorn==19.4.5
uwsgi==2.0.15
django-cors-headers==1.1.0
django-oauth-toolkit==1.0.0
djangorestframework==3.7.0
#mysqlclient==1.3.9
requests==2.18.1
django-jsonfield==1.0.1
django-filter==1.0.2
coreapi==2.3.3
celery==3.1.23
flower==0.9.1
redis==2.10.3

